Add files via upload
https://scratch.mit.edu/projects/855735030/

5/24/23 - Created "Original Code" version (original meaning code without assistance).
4/?/23 - Made asteroids change size and shape randomly.
3/8/23 - Added boss.
3/7/23 - Added new enemies.
3/4/23 - Added HP bubbles as a way to regain health points.
3/3/23 - Posted Game Demo. Still need to input spawns for ships.
---------------------------------------------------------------------------------
CREDITS:

_PhantomCoder_
The code and Non-Scratch Sprites by me.

Sandy_Spider
The prequel to this game, "Space Shooter III" was in my old account, Sandy_Spider.  I deleted that account, along with all the projects in it. I'm recreating them to be better using all that I have learned from Scratch.